refactor: standardize SolarWM entrypoint name
