+++ models/solarwm/solarwm.py
